Fix type casting for ticket price calculation in Ticketing.c

+++ IntroductionToC/BasicConcepts/Ticketing.c
@@ -7,11 +7,11 @@ int main() {
     // number of tickets ordered
     int tickets = 73;
     // price per ticket
-    float price = 7.45;
+    float price = (float) 7.45;
 
     // Task: calculate the total price and store it in a variable
     // Hint: use a float to store the result.
-    float totalTicketPrice = tickets * price;
+    float totalTicketPrice = (float) tickets * price;
 
     //Task: output the total in the format "Total: X", where X is the total price.
     printf("Total: %f\n", totalTicketPrice);
